* src/SDCC.y, src/SDCCsymt.h, src/SDCCutil.c, src/SDCCutil.h,   src/SDCCval.c: fixed flaws in previous commit
git-svn-id: http://svn.code.sf.net/p/sdcc/code/trunk@7465 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sdcc/src/SDCCsymt.h
@@ -215,7 +215,7 @@ DECLARATOR_TYPE;
 typedef struct declarator
 {
   DECLARATOR_TYPE dcl_type;         /* POINTER,ARRAY or FUNCTION  */
-  unsigned int num_elem;            /* # of elems if type==array, */
+  size_t num_elem;                  /* # of elems if type==array, */
   /* always 0 for flexible arrays */
   unsigned ptr_const:1;             /* pointer is constant        */
   unsigned ptr_volatile:1;          /* pointer is volatile        */
